+++ programs/redstone-sol/src/instructions/mod.rs
@@ -1,6 +1,3 @@
-pub mod constants;
 pub mod process_redstone_payload;
-pub mod redstone;
 
 pub use process_redstone_payload::*;
-pub use redstone::*;
