+++ programs/redstone-sol/src/lib.rs
@@ -1,9 +1,12 @@
+pub mod constants;
 pub mod error;
 pub mod instructions;
+pub mod redstone;
 pub mod state;
+pub mod util;
 
 use anchor_lang::prelude::*;
-use state::*;
+use instructions::*;
 
 declare_id!("rstK87maWhB9ywBwNEzV96rTqLvoUU6oXG3LGoTgyc4");
 
@@ -11,7 +14,10 @@ declare_id!("rstK87maWhB9ywBwNEzV96rTqLvoUU6oXG3LGoTgyc4");
 pub mod redstone_sol {
     use super::*;
 
-    pub fn process_redstone_payload(ctx: Context<ProcessPayload>, payload: Vec<u8>) -> Result<()> {
+    pub fn process_redstone_payload(
+        ctx: Context<ProcessPayload>,
+        payload: Vec<u8>,
+    ) -> Result<()> {
         instructions::process_redstone_payload(ctx, payload)
     }
 }
